% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_dt_faa_ntsb_clean_doc.R
\docType{data}
\name{dt_faa_ntsb_clean}
\alias{dt_faa_ntsb_clean}
\title{FAA aircraft data with NTSB event indicator flag}
\format{
\subsection{\code{dt_faa_ntsb_clean}}{

A data frame with 2,696,313 rows and 49 columns:
\describe{
\item{unique_id_line}{A unique id for each line of data}
\item{unique_id_ac}{A unique id for each aircraft.  Not related to the unique_id on the FAA data}
\item{n_number}{n number from the FAA data.  The prefix N has been added}
\item{serial_number}{serial number from the FAA data}
\item{source}{Source of the data; either the FAA master file or the FAA dergistered file}
\item{mfr_mdl_code}{Aircraft manufacture and model code from the FAA data}
\item{eng_mfr_mdl}{Engine manufacture and model code from the FAA data}
\item{unique_id_line}{A unique id for each line of data}
\describe{
\item{type_registrant}{The type of the registered owner. The codes and their corresponding descriptions are as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Code | Description                |
|------|----------------------------|
| 1    | Individual                 |
| 2    | Partnership                |
| 3    | Corporation                |
| 4    | Co-Owned                   |
| 5    | Government                 |
| 7    | LLC                        |
| 8    | Non Citizen Corporation    |
| 9    | Non Citizen Co-Owned       |
}\if{html}{\out{</div>}}

}

\item{street2_ind}{}
\item{city}{}
\item{state}{}
\item{zip_code}{}
\item{region}{}
\item{county}{}
\item{country}{}
\item{nu_registered}{}
\item{co_ownership}{}
\item{nu_n_number}{}
\item{co_owners_num}{}
\item{cert_issue_date}{}
\item{air_worth_date}{}
\item{expiration_date}{}
\item{last_action_date}{}
\item{airworthiness}{}
\item{operation}{}
\item{year_mfr}{}
\item{faa_acft_make}{}
\item{faa_acft_model}{}
\item{faa_acft_type_acft}{}
\item{faa_acft_type_eng}{}
\item{faa_acft_ac_cat}{}
\item{faa_acft_build_cert_ind}{}
\item{faa_acft_no_eng}{}
\item{faa_acft_no_seats}{}
\item{faa_acft_ac_weight}{}
\item{faa_acft_speed}{}
\item{faa_eng_make}{}
\item{faa_eng_model}{}
\item{faa_eng_type}{}
\item{faa_eng_hp}{}
\item{faa_eng_hp_char}{}
\item{faa_eng_thrust}{}
\item{faa_eng_thrust_char}{}
\item{kit_indyn}{}
\item{veh_age}{}
\item{start_date}{}
\item{end_date}{}
\item{ex}{}
\item{nu_cl}{}
}
}
\source{
\url{https://www.faa.gov/licenses_certificates/aircraft_certification/aircraft_registry/releasable_aircraft_download}
The NTSB keeps detailed records of every incidents in a access database.
The full data and the database structure can be downloaded from
avall.zip

\url{https://data.ntsb.gov/avdata/FileDirectory/DownloadFile?fileID=C\%3A\%5Cavdata\%5Cavall.zip}
codman.pdf

\url{https://data.ntsb.gov/avdata/FileDirectory/DownloadFile?fileID=C\%3A\%5Cavdata\%5Ccodman.pdf}
}
\usage{
dt_faa_ntsb_clean
}
\description{
The FAA maintains a list of registered aircraft with various details
including date the registration number (n number), the date aircraft was
certified and the expiration date.  The full data can be downloaded from
}
\keyword{datasets}
