% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_dt_faa_ntsb_clean_doc.R
\docType{data}
\name{dt_faa_ntsb_clean}
\alias{dt_faa_ntsb_clean}
\title{FAA aircraft data with NTSB event indicator flag}
\format{
\subsection{\code{dt_faa_ntsb_clean}}{

A data frame with 2,696,313 rows and 49 columns:

\describe{
\itemize{

\item{unique_id_line}{A unique id for each line of data}
\item{unique_id_ac}{A unique id for each aircraft.  Not related to the unique_id on the FAA data}
\item{n_number}{n number from the FAA data.  The prefix N has been added}
\item{serial_number}{serial number from the FAA data}
\item{source}{Source of the data; either the FAA master file or the FAA dergistered file}
\item{mfr_mdl_code}{Aircraft manufacture and model code from the FAA data}
\item{eng_mfr_mdl}{Engine manufacture and model code from the FAA data}
\item{type_registrant}{The type of the registered owner. The codes and their corresponding descriptions are as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Code | Description                |
|------|----------------------------|
| 1    | Individual                 |
| 2    | Partnership                |
| 3    | Corporation                |
| 4    | Co-Owned                   |
| 5    | Government                 |
| 7    | LLC                        |
| 8    | Non Citizen Corporation    |
| 9    | Non Citizen Co-Owned       |
}\if{html}{\out{</div>}}

}

\item{street2_ind}{0/1 indicator of whether or not there is second street line in address}
\item{city}{The city name which appears on the Application for Registration, AC Form 8050-1 or the latest address reported}
\item{state}{The state name which appears on the Application for Registration, AC Form 8050-1 or the latest address reported.}
\item{zip_code}{The postal Zip Code which appears on the Application for Registration, AC Form 8050-1 or the latest address reported.}

\item{region}{A character column indicating the region. The region codes are as follows (blanks in original data are replaced with X and may be derivable from city)
\itemize{
\item \code{1}: Eastern
\item \code{2}: Southwestern
\item \code{3}: Central
\item \code{4}: Western-Pacific
\item \code{5}: Alaskan
\item \code{7}: Southern
\item \code{8}: European
\item \code{C}: Great Lakes
\item \code{E}: New England
\item \code{S}: Northwest Mountain
} # close itemize
} # close item

\item{county}{A code representing the county which appears on the Application for Registration}
\item{country}{A code representing the country which appears on the Application for Registration}
\item{nu_registered}{A derived field: The number of aircraft registered to this owner}
\item{co_owners_num}{A derived field: The number of co-owners}
\item{co_ownership}{A derived field: Whether or not the aircraft is co-onwed}
\item{nu_n_number}{A derived field: The number of time an n_number appears in a data source}
\item{cert_issue_date}{}
\item{air_worth_date}{}
\item{expiration_date}{}
\item{last_action_date}{}

\item{airworthiness}{The airworthiness certificate class which is reported on the Application for Airworthiness, FAA Form 8130-6:
\itemize{
\item \code{1}: Standard
\item \code{2}: Limited
\item \code{3}: Restricted
\item \code{4}: Experimental
\item \code{5}: Provisional
\item \code{6}: Multiple
\item \code{7}: Primary
\item \code{8}: Special Flight Permit
\item \code{9}: Light Sport
}
}

\item{operation}{For some airworthiness codes there is more detail about the operation.  There can be up to 6 operations, only one is chosen:
\itemize{
\item \code{1}: Standard
\itemize{
\item \code{N}: Normal
\item \code{U}: Utility
\item \code{A}: Acrobatic
\item \code{T}: Transport
\item \code{G}: Glider
\item \code{B}: Balloon
\item \code{C}: Commuter
\item \code{X}: Other or blank
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{      \\item \code{2}: Limited
          \\itemize\{
               \\item \code{X}: No further detail within limited.  All coded to X
          \}
      \\item \code{3}: Restricted
          \\itemize\{
               \\item \code{0}: Other
               \\item \code{1}: Agriculture and Pest Control
               \\item \code{2}: Aerial Surveying
               \\item \code{3}: Aerial Advertising
               \\item \code{4}: Forest
               \\item \code{5}: Patrolling
               \\item \code{6}: Weather Control
               \\item \code{7}: Carriage of Cargo
          \}
      \\item \code{4}: Experimental
          \\itemize\{
               \\item \code{X}: Limited further detail in FAA so all coded to X
         \}
      \\item \code{5}: Provisional
          \\itemize\{
               \\item \code{X}: No further detail within Provisional  All coded to X
          \}
      \\item \code{6}: Multiple
          \\itemize\{
               \\item Same codes as code 3, restricted
          \}
      \\item \code{7}: Primary
          \\itemize\{
               \\item \code{X}: Very few of these (split between R&D and amateur) all coded to X
          \}
      \\item \code{8}: Special Flight Permit
          \\itemize\{
               \\item \code{X}: Mostly ferry flight for repairs, alterations, maintenance or storage.  All coded to X
          \}
      \\item \code{9}: Light Sport
          \\itemize\{
               \\item \code{A}: airplane
               \\item \code{X}: includes: G - Glider; L - Lighter than Air; P - Power-Parachute; W- Weight-Shift-Control,
          \}
\} # close itemize of operation
}\if{html}{\out{</div>}}

} # close item {operation}

\item{year_mfr}{Year manufactured.  Often missing}

\item{faa_acft_make}{aircraft make. merged from ReleasableAircraft/ACFTREF.txt}

\item{faa_acft_model}{aircraft model. merged from ReleasableAircraft/ACFTREF.txt}

\item{faa_acft_type_acft}{aircraft type. merged from ReleasableAircraft/ACFTREF.txt
\itemize{
\item \code{1}: Glider
\item \code{2}: Balloon
\item \code{3}: Blimp/Dirigible
\item \code{4}: Fixed wing single engine
\item \code{5}: Fixed wing multi engine
\item \code{6}: Rotorcraft
\item \code{7}: Weight-shift-control
\item \code{8}: Powered Parachute
\item \code{9}: Gyroplane
\item \code{H}: Hybrid Lift
\item \code{O}: Other
}
}

\item{faa_acft_type_eng}{aircraft engine type merged from ReleasableAircraft/ACFTREF.txt
\itemize{
\item \code{0}: None
\item \code{1}: Reciprocating
\item \code{2}: Turbo-prop
\item \code{3}: Turbo-shaft
\item \code{4}: Turbo-jet
\item \code{5}: Turbo-fan
\item \code{6}: Ramjet
\item \code{7}: 2 Cycle
\item \code{8}: 4 Cycle
\item \code{9}: Unknown
\item \code{10}: Electric
\item \code{11}: Rotary
}
}

\item{faa_acft_ac_cat}{aircraft category merged from ReleasableAircraft/ACFTREF.txt
\itemize{
\item \code{1}: Land
\item \code{2}: Sea
\item \code{3}: Amphibian
}
}

\item{faa_acft_build_cert_ind}{aircraft buidler certification code merged from ReleasableAircraft/ACFTREF.txt
\itemize{
\item \code{0}: Type Certificated
\item \code{1}: Not Type Certificated
\item \code{2}: Light Sport
}
}

\item{faa_acft_no_eng}{number of engines merged from ReleasableAircraft/ACFTREF.txt}
\item{faa_acft_no_seats}{number of seats merged from ReleasableAircraft/ACFTREF.txt}

\item{faa_acft_ac_weight}{Aircraft maximum gross take off weight in pounds  merged from ReleasableAircraft/ACFTREF.txt
\itemize{
\item \code{CLASS 1}: Up to 12,499 lb
\item \code{CLASS 2}: 12,500 - 19,999 lb
\item \code{CLASS 3}: 20,000 and over lb
\item \code{CLASS 4}: UAV up to 55 lb
}
}

\item{faa_acft_speed}{Aircraft average cruising speed in miles per hour. This data element is not present on all records.}

\item{faa_eng_make}{The name of the engine manufacturer. merged from ReleasableAircraft/ENGINE.txt}

\item{faa_eng_model}{The name of the engine model. merged from ReleasableAircraft/ENGINE.txt}
\item{faa_eng_type}{engine type merged from ReleasableAircraft/ENGINE.txt}
\itemize{
\item \code{0}: None
\item \code{1}: Reciprocating
\item \code{2}: Turbo-prop
\item \code{3}: Turbo-shaft
\item \code{4}: Turbo-jet
\item \code{5}: Turbo-fan
\item \code{6}: Ramjet
\item \code{7}: 2 Cycle
\item \code{8}: 4 Cycle
\item \code{9}: Unknown
\item \code{10}: Electric
\item \code{11}: Rotary
}
}
\item{faa_eng_hp}{Horsepower for engines types 1, 2, 3, 7, 8}
\item{faa_eng_hp_char}{}

\item{faa_eng_thrust}{Thrust for engine types 4, 5, 6}
\item{faa_eng_thrust_char}{}

\item{kit_indyn}{Y if kit manucaturer is not blank else N}
\item{veh_age}{age of aircraft based on year(start_date) - year(cert_issue_date)}
\item{start_date}{start date for exposure on each line}
\item{end_date}{end date of exposure on each line}
\item{ex}{pmin(1, as.integer(end_date - start_date + 1) / 365.25)}
\item{nu_cl}{number of incidents on the NTSB file between start and end date}
} # end itemize
} # end describe
}
}
\source{
\url{https://www.faa.gov/licenses_certificates/aircraft_certification/aircraft_registry/releasable_aircraft_download}
The NTSB keeps detailed records of every incidents in a access database.
The full data and the database structure can be downloaded from
avall.zip

\url{https://data.ntsb.gov/avdata/FileDirectory/DownloadFile?fileID=C\%3A\%5Cavdata\%5Cavall.zip}
codman.pdf

\url{https://data.ntsb.gov/avdata/FileDirectory/DownloadFile?fileID=C\%3A\%5Cavdata\%5Ccodman.pdf}
}
\usage{
dt_faa_ntsb_clean
}
\description{
The FAA maintains a list of registered aircraft with various details
including date the registration number (n number), the date aircraft was
certified and the expiration date.  The full data can be downloaded from
}
\keyword{datasets}
